scale-ai/training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
pika>=1.3.0
minio>=7.2.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
